@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -27,19 +27,19 @@ dependencies:
     sdk: flutter
   auto_size_text: 3.0.0
   cached_network_image: 3.2.1
-  cloud_firestore: 4.2.0
-  cloud_firestore_platform_interface: 5.9.1
-  cloud_firestore_web: 3.1.1
+  cloud_firestore: 4.8.0
+  cloud_firestore_platform_interface: 5.15.0
+  cloud_firestore_web: 3.6.0
   file_picker: 5.2.6
-  firebase_auth: 4.2.0
-  firebase_auth_platform_interface: 6.11.4
-  firebase_auth_web: 5.2.0
-  firebase_core: 2.4.0
-  firebase_core_platform_interface: 4.5.2
-  firebase_core_web: 2.0.2
-  firebase_storage: 11.0.7
-  firebase_storage_platform_interface: 4.1.25
-  firebase_storage_web: 3.3.17
+  firebase_auth: 4.6.2
+  firebase_auth_platform_interface: 6.15.2
+  firebase_auth_web: 5.5.2
+  firebase_core: 2.13.1
+  firebase_core_platform_interface: 4.8.0
+  firebase_core_web: 2.5.0
+  firebase_storage: 11.2.2
+  firebase_storage_platform_interface: 4.4.2
+  firebase_storage_web: 3.6.2
   flutter_animate: 4.1.1+1
   flutter_cache_manager: 3.3.0
   flutter_plugin_android_lifecycle: 2.0.9
@@ -51,7 +51,6 @@ dependencies:
   google_maps_flutter: 2.1.8
   google_maps_flutter_platform_interface: 2.2.1
   google_maps_flutter_web: 0.4.0+1
-  google_nav_bar: 5.0.5
   google_sign_in: 6.0.2
   google_sign_in_android: 6.1.8
   google_sign_in_ios: 5.6.1
@@ -62,7 +61,7 @@ dependencies:
   image_picker_for_web: 2.1.10
   image_picker_ios: 0.8.6+1
   image_picker_platform_interface: 2.6.2
-  intl: 0.17.0
+  intl: 0.18.0
   json_path: 0.4.1
   mime_type: 1.0.0
   page_transition: 2.0.4
@@ -135,6 +134,18 @@ flutter:
     - assets/rive_animations/
     - assets/pdfs/
 
+  fonts:
+    - family: 'dreqwdfa'
+      fonts:
+        - asset: assets/fonts/Simply_Rounded_Italic.ttf
+          style: italic
+        - asset: assets/fonts/Simply_Rounded.ttf
+          weight: 600
+        - asset: assets/fonts/Simply_Rounded_Bold.ttf
+          weight: 700
+        - asset: assets/fonts/Simply_Rounded_Bold_Italic.ttf
+          weight: 700
+          style: italic
 
 
   # An image asset can refer to one or more resolution-specific "variants", see
