@@ -44,10 +44,19 @@ T? castToType<T>(dynamic value) {
   }
   switch (T) {
     case double:
+      // Doubles may be stored as ints in some cases.
       return value.toDouble() as T;
+    case int:
+      // Likewise, ints may be stored as doubles. If this is the case
+      // (i.e. no decimal value), return the value as an int.
+      if (value is num && value.toInt() == value) {
+        return value.toInt() as T;
+      }
+      break;
     default:
-      return value as T;
+      break;
   }
+  return value as T;
 }
 
 extension MapDataExtensions on Map<String, dynamic> {
